Fix docs workspace build: textual -> rusty-textual rename fallout
- 27 docs/examples manifests: textual dep now names the rusty-textual
  package explicitly, keeping the short textual:: import ergonomics
- textual-macros: expansions rewrite hardcoded rusty_textual:: paths to
  the consumer's actual crate name (proc-macro-crate, entry-point
  retarget in new crate_path module); no-op for canonical/in-crate use
- Clear macro lints (if_same_then_else, unnecessary_unwrap,
  collapsible_if) and docs example warnings (dead_code, must_use,
  approx_constant); watcher &String signatures are framework-dictated
  (derive passes &FieldTy), allowed with reason
- Un-breaks tests/click_actions_pty, which builds docs/examples in setup

+++ docs/examples/guide/widgets/examples/counter02/main.rs
@@ -61,7 +61,7 @@ impl Widget for Counter {
         use rich_rs::{Segment, Segments};
         let text = format!("Count: {}", self.count);
         // Pad to widget width so the background fills the whole row.
-        let width = options.size.0 as usize;
+        let width = options.size.0;
         let padded = format!("{:<width$}", text, width = width);
         Segments::from(vec![Segment::new(padded)])
     }
